requirements.txt
fastapi
uvicorn
python-dotenv
requests
jinja2
python-jose
passlib
python-multipart
sqlalchemy
pydantic
pydantic-settings
aiofiles
python-dateutil
pyjwt
cryptography
pydantic[email]
bcrypt